Keep all inputs when importing a xml session
The Input.equals() relied on the dbId, which we don't have at this
point. The previous implementation was done to remove duplicates coming
from the DB. The collection of job inputs in ORM has been since then
changed from List to Set, so let's hope that takes care of the earlier
duplication problem.