Store job resource limits to session-db
Both scheduler and comp need this information. Scheduler uses it to set
container limits, comp uses it for variables like chipster.threads.max.
Storing it to db is the easiest way to allow it to be configured only in
one place.

Also more status updates were added for the client.